Check HF_TOKEN when logging
To support the intended use case of a public Space that only the
authorized users can write logs to (but anyone can view), pass in
and check the ownership of the HF_TOKEN against the owner of the
Space. Note: this only checks the ownership of the token against
the ownership of the Space, and does not look at any specific
permissions.